+++ react-app/pomodoro-timer-ui/src/Components/App/App.jsx
@@ -3,6 +3,7 @@ import Timer from "../Timer/Timer";
 import SidebarContainer from "../Sidebar/Sidebar";
 import './App.css'
 
+
 export default function App() {
   const [session, setSession] = React.useState("pomozone");
   return (
