+++ react-app/pomodoro-timer-ui/src/Components/ProfileTab/ProfileTab.css
@@ -1,12 +1,1 @@
-.profile-tab {
-    position: absolute;
-    width: 326px;
-    height: 372px;
-    left: 1654px;
-    top: 115px;
-
-    background: #F1F1F1;
-    border: 1px solid #000000;
-    border-radius: 10px;
-}
